Added first version of macros plugin.
The plugin is a formatter allowing to define macros (by means of regexp) and
expanse them inside any page. This, for instance, allows to easily create
shortcuts to chunk of code in which small part varies (one degree forward
simple variable offer by HTML::Template).

There is still some work to be done: simplification of macros definitions
which should be defined only headers.